Adds basic UIDynamics behaviour
We connect the cells to springs and then stretch them a distance proportional
to their distance from the touch point. Need to refine some values - and
consider whether we should try a full grid instead.

Also want to work on the gravity drop and bounce for inserting a new
view.